refactor(core): extract createCopyOrchestrator (PLAN §3.1.c, big completion)
The copy primitives + high-level entry points + the per-element label
copy variant — 5 inline functions totaling ~200 lines — pull out of
init() into a focused factory:

  - executeCopyOperation (cooldown clear, startCopy, label updates,
    completeCopy or deactivate or unfreeze)
  - copyResolvedElements (calls runCopyFlow with plugin hooks)
  - copyElementsToClipboard (per-element onElementSelect intercept +
    grabbed-box flash + notify-elements-selected event)
  - runCopyJob (the shared async tail used by both wrapper flows)
  - performCopyWithLabel (single-label entry)
  - performCopyWithPerElementLabels (per-element-label entry)
  - copyResolvedElements public re-export for api.copyElement

The CopyWithLabelOptions interface was previously defined inline in
init.tsx; moved next to the controller that owns it.

The factory takes (grab, pluginRegistry, labelManager,
copyFeedbackCooldown, deactivateRenderer callback). All other state
flows through those handles.

Drops 5 now-unused imports from index.tsx (CopyFailedError,
waitUntilNextFrame, getNearestComponentName, runCopyFlow,
notifyElementsSelected) plus the redundant
createPerElementLabelInstances/showTemporaryGrabbedBox destructures.

index.tsx: 2637 -> 2420 lines (-217 / cumulative -1304 / -35%).
40 copy-feedback + shift-multi-select specs green.
