+++ packages/react-grab/src/core/index.tsx
@@ -28,24 +28,21 @@ import { createEnterBlocker } from "./enter-blocker.js";
 import { createRendererHost } from "./renderer-host.js";
 import { createOverlayVisibility } from "./overlay-visibility.js";
 import { createPluginStateBridge } from "./plugin-state-bridge.js";
-import { CopyFailedError } from "../errors.js";
+import { createCopyOrchestrator } from "./copy-orchestrator.js";
 import {
   isKeyboardEventTriggeredByInput,
   hasTextSelectionInInput,
   hasTextSelectionOnPage,
 } from "../utils/is-keyboard-event-triggered-by-input.js";
 import { createComponentNameForElement } from "../utils/create-component-name-for-element.js";
-import { waitUntilNextFrame } from "../utils/native-raf.js";
 import {
   getStackContext,
-  getNearestComponentName,
   getComponentDisplayName,
   checkIsNextProject,
   resolveSource,
 } from "./context.js";
 import { createNoopApi } from "./noop-api.js";
 import { createEventListenerManager } from "./events.js";
-import { runCopyFlow } from "./copy.js";
 import {
   clearElementPositionCache,
   getElementAtPosition,
@@ -59,7 +56,6 @@ import { getElementAnchorRatio } from "../utils/get-element-anchor-ratio.js";
 import { createElementBounds } from "../utils/create-element-bounds.js";
 import { normalizeErrorMessage } from "../utils/normalize-error.js";
 import {
-  createBoundsFromDragRect,
   createFlatOverlayBounds,
   createPageRectFromBounds,
 } from "../utils/create-bounds-from-drag-rect.js";
@@ -118,25 +114,9 @@ import { commentPlugin } from "./plugins/comment.js";
 import { openPlugin } from "./plugins/open.js";
 import { freezeAllAnimations } from "../utils/freeze-animations.js";
 import { copyContent } from "../utils/copy-content.js";
-import { notifyElementsSelected } from "../utils/notify-elements-selected.js";
 import { logRecoverableError } from "../utils/log-recoverable-error.js";
 
 const builtInPlugins = [copyPlugin, commentPlugin, openPlugin];
-
-interface CopyWithLabelOptions {
-  element: Element;
-  cursorX: number;
-  selectedElements?: Element[];
-  extraPrompt?: string;
-  shouldDeactivateAfter?: boolean;
-  onComplete?: () => void;
-  dragRect?: {
-    pageX: number;
-    pageY: number;
-    width: number;
-    height: number;
-  };
-}
 
 interface BuildActionContextOptions {
   element: Element;
@@ -191,7 +171,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       theme: DEFAULT_THEME,
       keyHoldDuration: pluginRegistry.store.options.keyHoldDuration ?? DEFAULT_KEY_HOLD_DURATION_MS,
     });
-    const { store, actions, pointer, viewportVersion, current } = grab;
+    const { store, actions, pointer, viewportVersion } = grab;
 
     const phase = createGrabPhaseSelectors(grab);
     const {
@@ -225,11 +205,9 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       isGrabbedBoxesThemeEnabled: () => Boolean(pluginRegistry.store.theme.grabbedBoxes.enabled),
     });
     const {
-      showTemporaryGrabbedBox,
       createLabelInstance,
-      createPerElementLabelInstances,
-      clearAllLabels,
       updateLabelAfterCopy,
+      clearAllLabels,
       handleLabelInstanceHoverChange,
       computedLabelInstances,
       computedGrabbedBoxes,
@@ -342,7 +320,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
     let lastWindowFocusTimestamp = 0;
     const copyFeedbackCooldown = createCopyFeedbackCooldown();
     const clearCopyFeedbackCooldown = copyFeedbackCooldown.clear;
-    const startCopyFeedbackCooldown = copyFeedbackCooldown.start;
     let keyboardSelectedElement: Element | null = null;
     let pendingDefaultActionId: string | null = null;
     const [isPendingContextMenuSelect, setIsPendingContextMenuSelect] = createSignal(false);
@@ -369,212 +346,18 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
     );
 
 
-    const executeCopyOperation = async (
-      clipboardOperation: () => Promise<void>,
-      labelInstanceIds: string[] | null,
-      copiedElement?: Element,
-      shouldDeactivateAfter?: boolean,
-    ) => {
-      clearCopyFeedbackCooldown();
-      if (current().state !== "copying") {
-        actions.startCopy();
-      }
-
-      let didSucceed = false;
-      let errorMessage: string | undefined;
-
-      try {
-        await clipboardOperation();
-        didSucceed = true;
-      } catch (error) {
-        errorMessage = normalizeErrorMessage(error, "Action failed");
-      }
-
-      if (labelInstanceIds) {
-        for (const labelInstanceId of labelInstanceIds) {
-          updateLabelAfterCopy(labelInstanceId, didSucceed, errorMessage);
-        }
-      }
-
-      if (current().state !== "copying") return;
-
-      if (didSucceed) {
-        actions.completeCopy(copiedElement);
-      }
-
-      if (shouldDeactivateAfter) {
-        deactivateRenderer();
-      } else if (didSucceed) {
-        actions.activate();
-        startCopyFeedbackCooldown();
-      } else {
-        actions.unfreeze();
-      }
-    };
-
-    const copyResolvedElements = (
-      elements: Element[],
-      extraPrompt?: string,
-      resolvedComponentName?: string,
-    ) => {
-      const firstElement = elements[0];
-      const componentName =
-        resolvedComponentName ?? (firstElement ? getComponentDisplayName(firstElement) : null);
-      const tagName = firstElement ? getTagName(firstElement) : null;
-      const elementName = componentName ?? tagName ?? undefined;
-
-      return runCopyFlow(
-        {
-          getContent: pluginRegistry.store.options.getContent,
-          componentName: elementName,
-        },
-        {
-          onBeforeCopy: pluginRegistry.hooks.onBeforeCopy,
-          transformCopyContent: pluginRegistry.hooks.transformCopyContent,
-          onAfterCopy: pluginRegistry.hooks.onAfterCopy,
-          onCopySuccess: pluginRegistry.hooks.onCopySuccess,
-          onCopyError: pluginRegistry.hooks.onCopyError,
-        },
-        elements,
-        extraPrompt,
-      );
-    };
-
-    const copyElementsToClipboard = async (
-      targetElements: Element[],
-      extraPrompt?: string,
-      resolvedComponentName?: string,
-    ): Promise<void> => {
-      if (targetElements.length === 0) return;
-
-      const unhandledElements: Element[] = [];
-      const pendingResults: Promise<boolean>[] = [];
-      for (const element of targetElements) {
-        const { wasIntercepted, pendingResult } = pluginRegistry.hooks.onElementSelect(element);
-        if (!wasIntercepted) {
-          unhandledElements.push(element);
-        }
-        if (pendingResult) {
-          pendingResults.push(pendingResult);
-        }
-        if (pluginRegistry.store.theme.grabbedBoxes.enabled) {
-          showTemporaryGrabbedBox(createElementBounds(element), element);
-        }
-      }
-      await waitUntilNextFrame();
-      if (unhandledElements.length > 0) {
-        await copyResolvedElements(unhandledElements, extraPrompt, resolvedComponentName);
-      } else if (pendingResults.length > 0) {
-        const results = await Promise.all(pendingResults);
-        if (!results.every(Boolean)) {
-          throw new CopyFailedError();
-        }
-      }
-      void notifyElementsSelected(targetElements);
-    };
-
-    const runCopyJob = (job: {
-      primaryElement: Element;
-      elements: Element[];
-      labelInstanceIds: readonly string[];
-      extraPrompt?: string;
-      shouldDeactivateAfter?: boolean;
-      onComplete?: () => void;
-    }) => {
-      void getNearestComponentName(job.primaryElement)
-        .then(async (componentName) => {
-          await executeCopyOperation(
-            () =>
-              copyElementsToClipboard(job.elements, job.extraPrompt, componentName ?? undefined),
-            job.labelInstanceIds.length > 0 ? [...job.labelInstanceIds] : null,
-            job.primaryElement,
-            job.shouldDeactivateAfter,
-          );
-          job.onComplete?.();
-        })
-        .catch((error) => {
-          logRecoverableError("Copy operation failed", error);
-          const normalizedMessage = normalizeErrorMessage(error, "Action failed");
-          for (const labelInstanceId of job.labelInstanceIds) {
-            updateLabelAfterCopy(labelInstanceId, false, normalizedMessage);
-          }
-          if (current().state === "copying") {
-            actions.unfreeze();
-          }
-        });
-    };
-
-    const performCopyWithLabel = (options: CopyWithLabelOptions) => {
-      const {
-        element,
-        cursorX,
-        selectedElements,
-        extraPrompt,
-        shouldDeactivateAfter,
-        onComplete,
-        dragRect: passedDragRect,
-      } = options;
-
-      const allTargetElements = selectedElements ?? [element];
-      const dragRect = passedDragRect ?? store.frozenDragRect;
-      const isMultiSelect = allTargetElements.length > 1;
-
-      const selectionBounds =
-        dragRect && isMultiSelect
-          ? createBoundsFromDragRect(dragRect)
-          : createElementBounds(element);
-
-      const labelCursorX = isMultiSelect ? selectionBounds.x + selectionBounds.width / 2 : cursorX;
-
-      const tagName = getTagName(element);
-      clearCopyFeedbackCooldown();
-      actions.startCopy();
-
-      const labelInstanceId = tagName
-        ? createLabelInstance(selectionBounds, tagName, undefined, "copying", {
-            element,
-            mouseX: labelCursorX,
-            elements: selectedElements,
-          })
-        : null;
-
-      runCopyJob({
-        primaryElement: element,
-        elements: allTargetElements,
-        labelInstanceIds: labelInstanceId ? [labelInstanceId] : [],
-        extraPrompt,
-        shouldDeactivateAfter,
-        onComplete,
-      });
-    };
-
-    const performCopyWithPerElementLabels = (options: {
-      elements: Element[];
-      labelEntries: Array<{
-        element: Element;
-        tagName: string;
-        componentName?: string;
-        mouseX?: number;
-      }>;
-      shouldDeactivateAfter?: boolean;
-      onComplete?: () => void;
-    }) => {
-      const { elements, labelEntries, shouldDeactivateAfter, onComplete } = options;
-      const primaryElement = elements[0];
-
-      clearCopyFeedbackCooldown();
-      actions.startCopy();
-
-      const labelInstanceIds = createPerElementLabelInstances(labelEntries, "copying");
-
-      runCopyJob({
-        primaryElement,
-        elements,
-        labelInstanceIds,
-        shouldDeactivateAfter,
-        onComplete,
-      });
-    };
+    const copyOrchestrator = createCopyOrchestrator({
+      grab,
+      pluginRegistry,
+      labelManager,
+      copyFeedbackCooldown,
+      deactivateRenderer: () => deactivateRenderer(),
+    });
+    const {
+      performCopyWithLabel,
+      performCopyWithPerElementLabels,
+      copyResolvedElements,
+    } = copyOrchestrator;
 
     createOverlayEffects({
       grab,
